Move x86_64-specific code to arch/x86_64
