+++ src/arch/x86_64/interrupt/trace.rs
