+++ src/arch/x86_64/stop.rs
