+++ src/lib.rs
@@ -52,11 +52,13 @@ pub mod consts;
 /// ACPI table parsing
 mod acpi;
 
+/// Architecture-dependent stuff
+mod arch;
+#[cfg(target_arch = "x86_64")]
+pub use arch::x86_64::*;
+
 /// Context management
 pub mod context;
-
-/// Devices
-pub mod device;
 
 /// ELF file parsing
 pub mod elf;
@@ -64,32 +66,14 @@ pub mod elf;
 /// External functions
 pub mod externs;
 
-/// Global descriptor table
-pub mod gdt;
-
-/// Interrupt descriptor table
-mod idt;
-
-/// Interrupt instructions
-pub mod interrupt;
-
 /// Memory management
 pub mod memory;
-
-/// Paging
-pub mod paging;
 
 /// Panic
 pub mod panic;
 
 /// Schemes, filesystem handlers
 pub mod scheme;
-
-/// Initialization and start function
-pub mod start;
-
-/// Shutdown function
-pub mod stop;
 
 /// Synchronization primitives
 pub mod sync;
